Fixing pik encoder to a specific commit
Turns out they have migrated to usng Clang and whatever version or tools they
are using are currently very broken under CentOS 7. So until I have the time and
the willpower to go through and figure out why it's failing to build we'll just
run our harness on the last known good commit. At first glance it does not
appear to have any commits from this point in time to the last known state that
would impact codec performance - all commits appear to be more rearranging code
and the likes.